backend-mdr/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
gunicorn==21.2.0